move influxq_client package
